% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_splice_sites.R
\name{calculate_splice_sites}
\alias{calculate_splice_sites}
\title{Calculate Splice Sites}
\usage{
calculate_splice_sites(gtf_file)
}
\arguments{
\item{gtf_file}{A file path of GTF/GFF file}
}
\value{
A GRange object of splicing sites
}
\description{
This function can calculate splicing sites
}
