% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMAnno.R
\name{.AnnoColumn}
\alias{.AnnoColumn}
\title{Annotating gene information}
\usage{
.AnnoColumn(grange_file, selected_gr, GTF_type)
}
\arguments{
\item{grange_file}{Sites/Peak information file in GRange format}

\item{selected_gr}{Gene annotation information in GRange format}

\item{GTF_type}{Which database of GTF/GFF file from}
}
\value{
A GRange file annotated with the gene id, gene name and biotype.
}
\description{
This function can annotate basic gene information, including Gene_id,
Gene_name and Bio_type
}
