% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMAnno.R
\name{.AnnoIntergenic}
\alias{.AnnoIntergenic}
\title{Annotating intergenic region}
\usage{
.AnnoIntergenic(column_name, grange_file, selected_gr)
}
\arguments{
\item{column_name}{Column name for region information}

\item{grange_file}{Sites/Peak information file in GRange format}

\item{selected_gr}{Region annotation information in GRange format}
}
\value{
A GRange file annotated with the Region column
}
\description{
This function can annotate intergenic region information.
}
