% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMAnno.R
\name{.AnnoType}
\alias{.AnnoType}
\title{Annotating region information}
\usage{
.AnnoType(anno_info, column_name, grange_file, selected_gr)
}
\arguments{
\item{anno_info}{Region type (5'UTR, 3'UTR, Intron or Exon)}

\item{column_name}{Column name for region information}

\item{grange_file}{Sites/Peak information file in GRange format}

\item{selected_gr}{Region annotation information in GRange format}
}
\value{
A GRange file annotated with the Region column
}
\description{
This function can annotate region information, including 5'UTR, 3'UTR, Intron
and Exon.
}
