% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractGTFFile.R
\name{ExtractGTFFile}
\alias{ExtractGTFFile}
\title{Extract genomic information from GTF/GFF file.}
\usage{
ExtractGTFFile(gtf)
}
\arguments{
\item{gtf}{A file path of GTF/GFF file}
}
\value{
A list which contain genomic information
}
\description{
ExtractGTFFile could prepare the genomic information for the annotation step.
It extracts the region information including 5'UTR, 3'UTR, exon and intron
information. Moreover, it could capture the information of genes including
gene name, gene id and biotype if GTF/GFF files have that column.
}
