% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMAnno.R
\name{infoAnno}
\alias{infoAnno}
\title{Annotating other information}
\usage{
infoAnno(grange_file, other_gr, columns)
}
\arguments{
\item{grange_file}{Sites/Peak information file in GRange format}

\item{other_gr}{The grange object contain other information}

\item{columns}{The colnames that you want to annotate from other_gr}
}
\value{
A GRange file annotated with the other information.
}
\description{
This function can annotate other information, including RBP,
miRNA and SNP, which you need to provide the grange object for function.
}
