% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMAnno.R
\name{rmAnno}
\alias{rmAnno}
\title{Gene annotation}
\usage{
rmAnno(
  grange_file,
  gr_list,
  region = c("5'UTR", "3'UTR", "Exon", "Intron", "Intergenic")
)
}
\arguments{
\item{grange_file}{Sites/Peak information file in GRange format}

\item{gr_list}{A list contains genomic information extracted by ExtractGTFFile}

\item{region}{The vector of gene regions}
}
\value{
A annotated GRange file
}
\description{
This function can perform corresponding gene annotations according to the
site/peak information and GTF files provided by users. Annotation information
includes region, gene id, gene name and biotype information.
}
