% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_Distribution_gene.R
\name{site_Distribution_gene}
\alias{site_Distribution_gene}
\title{Visualization of RNA modification sites}
\usage{
site_Distribution_gene(
  gtf_file,
  geneName,
  seqInfo,
  gtf_path,
  gr1,
  mergeSample = F,
  rightmargin = 0.2,
  ideogramheight = 5,
  data1height = 200,
  gene_r0 = 0,
  gene_r1 = 0.6,
  mergeTranscripts = T,
  non.coding.exons.height = 0.5,
  non.coding.exons.col = "#82BB22",
  non.coding.exons.border.col = "#000000",
  coding.exons.col = "#0050A4",
  coding.exons.border.col = "#000000",
  introns.col = "#000000",
  region_col = "#E50053",
  region_r1 = 0.8,
  region_r0 = 0,
  point1_col = "#E50053",
  point1_r1 = 0.8,
  point1_r0 = 0.6,
  point1_pch = 16,
  point1_cex = 2,
  point1_y = 1,
  point2_col = "#000000",
  point2_r1 = 0.8,
  point2_r0 = 0.6,
  point2_pch = 1,
  point2_cex = 2,
  point2_y = 1,
  kpPoints = T
)
}
\arguments{
\item{gtf_file}{A data.frame or GRanges object containing gene annotations in GTF format.}

\item{geneName}{Character string specifying the gene symbol to plot. Only one gene can be visualized at a time.}

\item{seqInfo}{A Seqinfo object containing chromosome length information. If NULL, chromosome lengths will be automatically extracted from the GTF file.}

\item{gtf_path}{Character string specifying the path to a GTF file. Ignored if gtf_file is provided.}

\item{gr1}{A GRanges object containing genomic positions to be plotted as data points or regions.}

\item{mergeSample}{Logical indicating whether to merge multiple samples when plotting. If TRUE, data from different samples will be aggregated.}

\item{rightmargin}{Numeric value specifying the right margin of the plot as a fraction of the total plot width. Default is 0.1 10.}

\item{ideogramheight}{Numeric value between 0 and 1 specifying the height of the chromosome ideogram track relative to the total plot height.}

\item{data1height}{Numeric value between 0 and 1 specifying the height of the main data track relative to the total plot height.}

\item{gene_r0}{Numeric values between 0 and 1 defining the radial position inner and outer radius of the gene track in circular plots. gene_r0 is the inner radius, gene_r1 is the outer radius.}

\item{gene_r1}{Numeric values between 0 and 1 defining the radial position inner and outer radius of the gene track in circular plots. gene_r0 is the inner radius, gene_r1 is the outer radius.}

\item{mergeTranscripts}{Logical indicating whether to merge transcript variants of the same gene into a single representation.}

\item{non.coding.exons.height}{Numeric value specifying the height of non-coding exons in the gene model representation.}

\item{non.coding.exons.col}{Numeric value specifying the height of non-coding exons in the gene model representation.}

\item{non.coding.exons.border.col}{Character string specifying the fill color and border color for non-coding exons.}

\item{coding.exons.col}{Character string specifying the fill color and border color for non-coding exons.}

\item{coding.exons.border.col}{Character string specifying the  fill color and border color for coding exons.}

\item{introns.col}{Character string specifying the color for intron lines.}

\item{region_col}{Character string specifying the fill color for genomic regions.}

\item{region_r1}{Numeric values between 0 and 1 defining the radial position inner and outer radius of region tracks in circular plots.}

\item{region_r0}{Character string specifying the fill color for genomic regions.}

\item{point1_col}{Character string specifying the color for primary data points.}

\item{point1_r1}{Numeric values between 0 and 1 defining the radial position inner and outer radius for primary data points in circular plots.}

\item{point1_r0}{Numeric values between 0 and 1 defining the radial position inner and outer radius for primary data points in circular plots.}

\item{point1_pch}{Integer or single character specifying the plotting character for primary data points see par"pch".}

\item{point1_cex}{Numeric value specifying the character expansion factor for primary data points.}

\item{point1_y}{Numeric value or vector specifying the y-coordinates for primary data points in linear plots.}

\item{point2_col}{Character string specifying the color for secondary data points.}

\item{point2_r1}{Numeric values between 0 and 1 defining the radial position inner and outer radius for secondary data points in circular plots.}

\item{point2_r0}{Numeric values between 0 and 1 defining the radial position inner and outer radius for secondary data points in circular plots.}

\item{point2_pch}{Integer or single character specifying the plotting character for secondary data points see par"pch".}

\item{point2_cex}{Numeric value specifying the character expansion factor for secondary data points.}

\item{point2_y}{Numeric value or vector specifying the y-coordinates for secondary data points in linear plots.}

\item{kpPoints}{A function or expression specifying custom point plotting using karyoploteR::kpPoints. Useful for adding additional data layers.}
}
\description{
site_Distribution_gene visualize the RNA modification sites based on the
genomic region information and gene region information.
}
